requirements.txt
django==4.2.16
pdfplumber==0.11.4
anthropic
python-dotenv==1.0.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
whitenoise==6.7.0
gunicorn==22.0.0
Pillow==10.4.0
httpx==0.27.2
python-docx
striprtf
PyPDF2>=3.0.0
celery>=5.3.0
redis>=5.0.0